위키 문서 관련 기능을 하나의 컨트롤러로 통합
위키 문서와 관련된 기능들은 `Document` 컨트롤러에서 구현됩니다. 분리
되었던 위키 문서 작성 기능의 구현 내용이 이동되었습니다.
